feat: Tareas

+++ src/tasks/tasks.service.ts
@@ -1,7 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Task } from './entities/task.entity';
+import { Task } from './task.entity';
+import { CreateTaskDto, UpdateTaskDto } from './dto/taks.dto';
 
 @Injectable()
 export class TasksService {
@@ -18,17 +23,38 @@ export class TasksService {
     return this.tasksRepository.findOneBy({ id });
   }
 
-  create(task: Partial<Task>): Promise<Task> {
+  create(task: CreateTaskDto): Promise<Task> {
     const newTask = this.tasksRepository.create(task);
     return this.tasksRepository.save(newTask);
   }
 
-  async update(id: number, task: Partial<Task>): Promise<Task> {
-    await this.tasksRepository.update(id, task);
-    return this.findOne(id);
+  async update(id: number, updateTaskDto: UpdateTaskDto): Promise<Task> {
+    if (Object.keys(updateTaskDto).length === 0) {
+      throw new BadRequestException('No data submitted');
+    }
+
+    const task = await this.tasksRepository.findOne({ where: { id } });
+
+    if (!task) {
+      throw new NotFoundException(`Task with ID ${id} not found`);
+    }
+
+    Object.assign(task, updateTaskDto);
+
+    return this.tasksRepository.save(task);
   }
 
-  async remove(id: number): Promise<void> {
+  async remove(id: number) {
+    const task = await this.tasksRepository.findOne({ where: { id } });
+
+    if (!task) {
+      throw new NotFoundException(`Task with ID ${id} not found`);
+    }
+
     await this.tasksRepository.delete(id);
+
+    return {
+      message: `Task with ID ${id} deleted`,
+    };
   }
 }
